+++ code/src/sixsq/nuvla/server/resources/credential.clj
@@ -130,7 +130,7 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
 ;; actions
 ;;
 
-(defn create-job
+(defn create-check-credential-job
   [{{uuid :uuid} :params :as request}]
   (try
     (let [id (str resource-type "/" uuid)]
@@ -155,12 +155,29 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
   (let [id       (str resource-type "/" uuid)
         resource (crud/retrieve-by-id-as-admin id)]
     (a/throw-cannot-manage resource request)
-    (create-job request)))
+    (create-check-credential-job request)))
 
 
 ;;
 ;; CRUD operations
 ;;
+
+(def dispatch-by-first-arg-method
+  (fn [{:keys [method] :as _resource} _request] method))
+
+(defmulti post-add-hook dispatch-by-first-arg-method)
+
+(defmethod post-add-hook :default
+  [_resource _request]
+  nil)
+
+(defmulti post-edit-hook dispatch-by-first-arg-method)
+
+
+;; default post-add hook is a no-op
+(defmethod post-edit-hook :default
+  [_resource _request]
+  nil)
 
 (def add-impl (std-crud/add-fn resource-type collection-acl resource-type))
 
@@ -214,7 +231,7 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
   [{:keys [body] :as request}]
   (let [authn-info (auth/current-authentication request)
         desc-attrs (u/select-desc-keys body)
-        [create-resp {:keys [id] :as body}]
+        [create-resp body]
         (-> body
             (assoc :resource-type create-type)
             (update-in [:template] dissoc :subtype)         ;; forces use of template reference
@@ -225,16 +242,15 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
             (tpl->credential request))
 
         response   (-> request
-                       (assoc :id id :body (merge body desc-attrs))
+                       (assoc :body (merge body desc-attrs))
                        add-impl
                        (update-in [:body] merge create-resp))
 
-        id         (:resource-id (:body response))]
+        id         (-> response :body :resource-id)
+        cred       (assoc body :id id)]
 
-    (when (= (:method body) "infrastructure-service-swarm")
-      (create-job {:params      {:uuid          (u/id->uuid id)
-                                 :resource-name resource-type}
-                   :nuvla/authn authn-info}))
+    (post-add-hook cred request)
+
     response))
 
 
@@ -270,16 +286,18 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
 
 (defmethod crud/edit resource-type
   [{{uuid :uuid} :params body :body :as request}]
-  (let [subtype (-> (str resource-type "/" uuid)
-                    (db/retrieve request)
-                    :subtype)]
-    (-> body
-        (assoc :subtype subtype)
-        (dissoc :last-check)
-        (cond-> (:status body) (assoc :last-check (time/now-str)))
-        (special-edit request)
-        (->> (assoc request :body)
-             (edit-impl)))))
+  (let [subtype  (-> (str resource-type "/" uuid)
+                     (db/retrieve request)
+                     :subtype)
+        response (-> body
+                     (assoc :subtype subtype)
+                     (dissoc :last-check)
+                     (cond-> (:status body) (assoc :last-check (time/now-str)))
+                     (special-edit request)
+                     (->> (assoc request :body)
+                          (edit-impl)))]
+    (post-edit-hook (:body response) request)
+    response))
 
 
 (defn update-credential